requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gnupg==2.3.1
greenlet==1.1.2
ldap3==2.9.1
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
paramiko==2.11.0
PGPy==0.5.4
psutil==5.9.1
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyodbc==4.0.32
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
tzdata==2022.1
wfastcgi==3.0.0
